Reuse export generation checks in existing JSON Schema tests
